fix: 캐시 초기화를 막기 위해 queryClient를 상태로 관리
